fix(query): use Decimal for positions monetary fields and weight

+++ src/services/query_service/app/dtos/position_dto.py
@@ -1,5 +1,6 @@
 # services/query-service/app/dtos/position_dto.py
 from datetime import date
+from decimal import Decimal
 from typing import List, Optional
 
 from pydantic import BaseModel, ConfigDict, Field
@@ -33,10 +34,10 @@ class Position(BaseModel):
     country_of_risk: Optional[str] = Field(
         None, description="Instrument country of risk (ISO 3166-1 alpha-2).", examples=["US"]
     )
-    cost_basis: float = Field(
+    cost_basis: Decimal = Field(
         ..., description="Cost basis in portfolio base currency.", examples=[15000.0]
     )
-    cost_basis_local: Optional[float] = Field(
+    cost_basis_local: Optional[Decimal] = Field(
         None, description="Cost basis in local instrument currency.", examples=[15000.0]
     )
     valuation: Optional[ValuationData] = Field(
@@ -52,7 +53,7 @@ class Position(BaseModel):
         description="Start date of the current continuous holding period in the active epoch.",
         examples=["2025-01-15"],
     )
-    weight: Optional[float] = Field(
+    weight: Optional[Decimal] = Field(
         None,
         description="Position weight versus total portfolio market value (0.0 to 1.0).",
         examples=[0.2417],
@@ -78,11 +79,11 @@ class PositionHistoryRecord(BaseModel):
     )
     quantity: float = Field(..., description="The number of shares held as of this record.")
 
-    cost_basis: float = Field(
+    cost_basis: Decimal = Field(
         ..., description="The total cost basis of the holding as of this record."
     )
 
-    cost_basis_local: Optional[float] = Field(
+    cost_basis_local: Optional[Decimal] = Field(
         None, description="The total cost basis in the instrument's local currency."
     )
